+++ src/ldc/supervised/pairs/_alpaca.py
@@ -120,12 +120,14 @@ class AlpacaWriter(BatchPairWriter):
     Writer for the Alpaca JSON format.
     """
 
-    def __init__(self, target: str = None, logger_name: str = None, logging_level: str = LOGGING_WARN):
+    def __init__(self, target: str = None, pretty_print: bool = False, logger_name: str = None, logging_level: str = LOGGING_WARN):
         """
         Initializes the writer.
 
         :param target: the filename/dir to write to
         :type target: str
+        :param pretty_print: whether to output the JSON in more human-readable format
+        :type pretty_print: bool
         :param logger_name: the name to use for the logger
         :type logger_name: str
         :param logging_level: the logging level to use
@@ -133,6 +135,7 @@ class AlpacaWriter(BatchPairWriter):
         """
         super().__init__(logger_name=logger_name, logging_level=logging_level)
         self.target = target
+        self.pretty_print = pretty_print
         self._current_output = None
         self._output = None
 
@@ -163,6 +166,7 @@ class AlpacaWriter(BatchPairWriter):
         """
         parser = super()._create_argparser()
         parser.add_argument("-o", "--output", type=str, help="Path of the Alpaca file to write (directory when processing multiple files)", required=True)
+        parser.add_argument("-p", "--pretty_print", action="store_true", help="Whether to output the JSON in more human-readable format.", required=False)
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -174,6 +178,7 @@ class AlpacaWriter(BatchPairWriter):
         """
         super()._apply_args(ns)
         self.target = ns.output
+        self.pretty_print = ns.pretty_print
 
     def write_batch(self, data: Iterable[PairData]):
         """
@@ -189,10 +194,11 @@ class AlpacaWriter(BatchPairWriter):
             self._output = open_file(self._current_output, mode="wt")
 
         dicts = [x.to_dict() for x in data]
+        indent = 2 if self.pretty_print else None
         try:
-            json.dump(dicts, self._output, ensure_ascii=False)
+            json.dump(dicts, self._output, ensure_ascii=False, indent=indent)
         except:
-            json.dump(dicts, self._output, ensure_ascii=True)
+            json.dump(dicts, self._output, ensure_ascii=True, indent=indent)
 
     def finalize(self):
         """