+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 - added the `llm-registry` tool for querying the registry from the command-line
 - added the `replace_patterns` method to `ldc.text_utils` module
 - added the `replace-patterns` filter
+- added `-p/--pretty-print` flag to `to-alpaca` writer
 
 
 0.0.2 (2023-10-31)